requirements.txt
# need for discord integration
requests
pynacl

# other dependencies
boto3
